[WIP] slab

+++ lab6/include/list.h
@@ -1,7 +1,7 @@
 #pragma once
 
 struct Pair {
-  int lb, ub;
+  unsigned long lb, ub;
 };
 
 struct Node {
